+++ GUI/MainChangeset.cs
@@ -76,6 +76,18 @@ namespace CKAN
             }
         }
 
+        private void ClearChangeSet()
+        {
+            foreach (DataGridViewRow row in mainModList.full_list_of_mod_rows.Values)
+            {
+                GUIMod mod = row.Tag as GUIMod;
+                if (mod.IsInstallChecked != mod.IsInstalled)
+                {
+                    mod.SetInstallChecked(row, mod.IsInstalled);
+                }
+            }
+        }
+
         /// <summary>
         /// This method creates the Install part of the changeset
         /// It arranges the changeset in a human-friendly order
@@ -104,11 +116,9 @@ namespace CKAN
 
         private void CancelChangesButton_Click(object sender, EventArgs e)
         {
-            UpdateModsList();
+            ClearChangeSet();
             UpdateChangesDialog(null, installWorker);
             tabController.ShowTab("ManageModsTabPage");
-            tabController.HideTab("ChangesetTabPage");
-            ApplyToolButton.Enabled = false;
         }
 
         private void ConfirmChangesButton_Click(object sender, EventArgs e)
@@ -119,15 +129,13 @@ namespace CKAN
             menuStrip1.Enabled = false;
             RetryCurrentActionButton.Visible = false;
 
-            RelationshipResolverOptions install_ops = RelationshipResolver.DefaultOpts();
-            install_ops.with_recommends = false;
             //Using the changeset passed in can cause issues with versions.
             // An example is Mechjeb for FAR at 25/06/2015 with a 1.0.2 install.
             // TODO Work out why this is.
             installWorker.RunWorkerAsync(
                 new KeyValuePair<List<ModChange>, RelationshipResolverOptions>(
                     mainModList.ComputeUserChangeSet().ToList(),
-                    install_ops
+                    RelationshipResolver.DependsOnlyOpts()
                 )
             );
         }