+++ GUI/MainModInfo.Designer.cs
@@ -32,28 +32,38 @@
             this.MetadataTabPage = new System.Windows.Forms.TabPage();
             this.splitContainer2 = new System.Windows.Forms.SplitContainer();
             this.MetaDataUpperLayoutPanel = new System.Windows.Forms.TableLayoutPanel();
-            this.MetadataModuleNameLabel = new System.Windows.Forms.Label();
-            this.MetadataModuleAbstractLabel = new System.Windows.Forms.Label ();
-            this.MetadataModuleDescriptionLabel = new System.Windows.Forms.Label ();
+            this.MetadataModuleNameTextBox = new TransparentTextBox();
+            this.MetadataModuleAbstractLabel = new System.Windows.Forms.Label();
+            this.MetadataModuleDescriptionTextBox = new TransparentTextBox();
             this.MetaDataLowerLayoutPanel = new System.Windows.Forms.TableLayoutPanel();
             this.IdentifierLabel = new System.Windows.Forms.Label();
-            this.MetadataIdentifierLabel = new System.Windows.Forms.Label();
+            this.MetadataIdentifierTextBox = new TransparentTextBox();
             this.KSPCompatibilityLabel = new System.Windows.Forms.Label();
             this.ReleaseLabel = new System.Windows.Forms.Label();
             this.GitHubLabel = new System.Windows.Forms.Label();
             this.HomePageLabel = new System.Windows.Forms.Label();
             this.AuthorLabel = new System.Windows.Forms.Label();
             this.LicenseLabel = new System.Windows.Forms.Label();
-            this.MetadataModuleVersionLabel = new System.Windows.Forms.Label();
-            this.MetadataModuleLicenseLabel = new System.Windows.Forms.Label();
-            this.MetadataModuleAuthorLabel = new System.Windows.Forms.Label();
+            this.MetadataModuleVersionTextBox = new TransparentTextBox();
+            this.MetadataModuleLicenseTextBox = new TransparentTextBox();
+            this.MetadataModuleAuthorTextBox = new TransparentTextBox();
             this.VersionLabel = new System.Windows.Forms.Label();
-            this.MetadataModuleReleaseStatusLabel = new System.Windows.Forms.Label();
+            this.MetadataModuleReleaseStatusTextBox = new TransparentTextBox();
             this.MetadataModuleHomePageLinkLabel = new System.Windows.Forms.LinkLabel();
-            this.MetadataModuleKSPCompatibilityLabel = new System.Windows.Forms.Label();
+            this.MetadataModuleKSPCompatibilityTextBox = new TransparentTextBox();
             this.MetadataModuleGitHubLinkLabel = new System.Windows.Forms.LinkLabel();
             this.RelationshipTabPage = new System.Windows.Forms.TabPage();
             this.DependsGraphTree = new System.Windows.Forms.TreeView();
+            this.LegendDependsImage = new System.Windows.Forms.PictureBox();
+            this.LegendRecommendsImage = new System.Windows.Forms.PictureBox();
+            this.LegendSuggestsImage = new System.Windows.Forms.PictureBox();
+            this.LegendSupportsImage = new System.Windows.Forms.PictureBox();
+            this.LegendConflictsImage = new System.Windows.Forms.PictureBox();
+            this.LegendDependsLabel = new System.Windows.Forms.Label();
+            this.LegendRecommendsLabel = new System.Windows.Forms.Label();
+            this.LegendSuggestsLabel = new System.Windows.Forms.Label();
+            this.LegendSupportsLabel = new System.Windows.Forms.Label();
+            this.LegendConflictsLabel = new System.Windows.Forms.Label();
             this.ContentTabPage = new System.Windows.Forms.TabPage();
             this.ContentsPreviewTree = new System.Windows.Forms.TreeView();
             this.ContentsDownloadButton = new System.Windows.Forms.Button();
@@ -126,9 +136,9 @@
             this.MetaDataUpperLayoutPanel.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
             this.MetaDataUpperLayoutPanel.ColumnCount = 1;
             this.MetaDataUpperLayoutPanel.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100F));
-            this.MetaDataUpperLayoutPanel.Controls.Add(this.MetadataModuleNameLabel, 0, 0);
+            this.MetaDataUpperLayoutPanel.Controls.Add(this.MetadataModuleNameTextBox, 0, 0);
             this.MetaDataUpperLayoutPanel.Controls.Add(this.MetadataModuleAbstractLabel, 0, 1);
-            this.MetaDataUpperLayoutPanel.Controls.Add(this.MetadataModuleDescriptionLabel, 0, 2);
+            this.MetaDataUpperLayoutPanel.Controls.Add(this.MetadataModuleDescriptionTextBox, 0, 2);
             this.MetaDataUpperLayoutPanel.Dock = System.Windows.Forms.DockStyle.Fill;
             this.MetaDataUpperLayoutPanel.Location = new System.Drawing.Point(0, 0);
             this.MetaDataUpperLayoutPanel.Name = "MetaDataUpperLayoutPanel";
@@ -139,17 +149,20 @@
             this.MetaDataUpperLayoutPanel.Size = new System.Drawing.Size(346, 283);
             this.MetaDataUpperLayoutPanel.TabIndex = 0;
             //
-            // MetadataModuleNameLabel
+            // MetadataModuleNameTextBox
             //
-            this.MetadataModuleNameLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleNameLabel.Font = new System.Drawing.Font("Microsoft Sans Serif", 9.75F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.MetadataModuleNameLabel.ForeColor = System.Drawing.SystemColors.GrayText;
-            this.MetadataModuleNameLabel.Location = new System.Drawing.Point(3, 0);
-            this.MetadataModuleNameLabel.Name = "MetadataModuleNameLabel";
-            this.MetadataModuleNameLabel.Size = new System.Drawing.Size(340, 46);
-            this.MetadataModuleNameLabel.TabIndex = 0;
-            this.MetadataModuleNameLabel.Text = "Mod Name";
-            this.MetadataModuleNameLabel.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
+            this.MetadataModuleNameTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleNameTextBox.Font = new System.Drawing.Font("Microsoft Sans Serif", 9.75F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.MetadataModuleNameTextBox.Location = new System.Drawing.Point(3, 0);
+            this.MetadataModuleNameTextBox.Name = "MetadataModuleNameTextBox";
+            this.MetadataModuleNameTextBox.Size = new System.Drawing.Size(340, 46);
+            this.MetadataModuleNameTextBox.TabIndex = 0;
+            this.MetadataModuleNameTextBox.Text = "Mod Name";
+            this.MetadataModuleNameTextBox.TextAlign = System.Windows.Forms.HorizontalAlignment.Center;
+            this.MetadataModuleNameTextBox.ReadOnly = true;
+            this.MetadataModuleNameTextBox.BackColor = MetaDataUpperLayoutPanel.BackColor;
+            this.MetadataModuleNameTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleNameTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // MetadataModuleAbstractLabel
             //
@@ -158,19 +171,22 @@
             this.MetadataModuleAbstractLabel.Location = new System.Drawing.Point(3, 49);
             this.MetadataModuleAbstractLabel.Name = "MetadataModuleAbstractLabel";
             this.MetadataModuleAbstractLabel.Size = new System.Drawing.Size(340, 61);
+            this.MetadataModuleAbstractLabel.AutoSize = true;
             this.MetadataModuleAbstractLabel.TabIndex = 27;
             this.MetadataModuleAbstractLabel.Text = "";
             //
-            // MetadataModuleDescriptionLabel
+            // MetadataModuleDescriptionTextBox
             //
-            this.MetadataModuleDescriptionLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-
-            this.MetadataModuleDescriptionLabel.ForeColor = System.Drawing.SystemColors.ControlText;
-            this.MetadataModuleDescriptionLabel.Location = new System.Drawing.Point (3, 129);
-            this.MetadataModuleDescriptionLabel.Name = "MetadataModuleDescriptionLabel";
-            this.MetadataModuleDescriptionLabel.Size = new System.Drawing.Size (340, 121);
-            this.MetadataModuleDescriptionLabel.TabIndex = 28;
-            this.MetadataModuleDescriptionLabel.Text = "";
+            this.MetadataModuleDescriptionTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleDescriptionTextBox.Location = new System.Drawing.Point (3, 129);
+            this.MetadataModuleDescriptionTextBox.Name = "MetadataModuleDescriptionLabel";
+            this.MetadataModuleDescriptionTextBox.Size = new System.Drawing.Size (340, 121);
+            this.MetadataModuleDescriptionTextBox.TabIndex = 28;
+            this.MetadataModuleDescriptionTextBox.Text = "";
+            this.MetadataModuleDescriptionTextBox.ReadOnly = true;
+            this.MetadataModuleDescriptionTextBox.BackColor = MetaDataUpperLayoutPanel.BackColor;
+            this.MetadataModuleDescriptionTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleDescriptionTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // MetaDataLowerLayoutPanel
             //
@@ -186,14 +202,14 @@
             this.MetaDataLowerLayoutPanel.Controls.Add(this.ReleaseLabel, 0, 5);
             this.MetaDataLowerLayoutPanel.Controls.Add(this.KSPCompatibilityLabel, 0, 6);
             this.MetaDataLowerLayoutPanel.Controls.Add(this.IdentifierLabel, 0, 7);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleVersionLabel, 1, 0);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleLicenseLabel, 1, 1);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleAuthorLabel, 1, 2);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleVersionTextBox, 1, 0);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleLicenseTextBox, 1, 1);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleAuthorTextBox, 1, 2);
             this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleHomePageLinkLabel, 1, 3);
             this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleGitHubLinkLabel, 1, 4);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleReleaseStatusLabel, 1, 5);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleKSPCompatibilityLabel, 1, 6);
-            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataIdentifierLabel, 1, 7);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleReleaseStatusTextBox, 1, 5);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataModuleKSPCompatibilityTextBox, 1, 6);
+            this.MetaDataLowerLayoutPanel.Controls.Add(this.MetadataIdentifierTextBox, 1, 7);
             this.MetaDataLowerLayoutPanel.Dock = System.Windows.Forms.DockStyle.Fill;
             this.MetaDataLowerLayoutPanel.Location = new System.Drawing.Point(0, 0);
             this.MetaDataLowerLayoutPanel.Name = "MetaDataLowerLayoutPanel";
@@ -222,16 +238,19 @@
             this.IdentifierLabel.TabIndex = 28;
             this.IdentifierLabel.Text = "Identifier";
             //
-            // MetadataIdentifierLabel
+            // MetadataIdentifierTextBox
             //
-            this.MetadataIdentifierLabel.AutoSize = true;
-            this.MetadataIdentifierLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataIdentifierLabel.ForeColor = System.Drawing.SystemColors.ControlText;
-            this.MetadataIdentifierLabel.Location = new System.Drawing.Point(93, 210);
-            this.MetadataIdentifierLabel.Name = "MetadataIdentifierLabel";
-            this.MetadataIdentifierLabel.Size = new System.Drawing.Size(250, 20);
-            this.MetadataIdentifierLabel.TabIndex = 27;
-            this.MetadataIdentifierLabel.Text = "-";
+            this.MetadataIdentifierTextBox.AutoSize = true;
+            this.MetadataIdentifierTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataIdentifierTextBox.Location = new System.Drawing.Point(93, 210);
+            this.MetadataIdentifierTextBox.Name = "MetadataIdentifierTextBox";
+            this.MetadataIdentifierTextBox.Size = new System.Drawing.Size(250, 20);
+            this.MetadataIdentifierTextBox.TabIndex = 27;
+            this.MetadataIdentifierTextBox.Text = "-";
+            this.MetadataIdentifierTextBox.ReadOnly = true;
+            this.MetadataIdentifierTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataIdentifierTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataIdentifierTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // KSPCompatibilityLabel
             //
@@ -299,35 +318,47 @@
             this.LicenseLabel.TabIndex = 3;
             this.LicenseLabel.Text = "License:";
             //
-            // MetadataModuleVersionLabel
+            // MetadataModuleVersionTextBox
             //
-            this.MetadataModuleVersionLabel.AutoSize = true;
-            this.MetadataModuleVersionLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleVersionLabel.Location = new System.Drawing.Point(93, 0);
-            this.MetadataModuleVersionLabel.Name = "MetadataModuleVersionLabel";
-            this.MetadataModuleVersionLabel.Size = new System.Drawing.Size(250, 30);
-            this.MetadataModuleVersionLabel.TabIndex = 2;
-            this.MetadataModuleVersionLabel.Text = "0.0.0";
+            this.MetadataModuleVersionTextBox.AutoSize = true;
+            this.MetadataModuleVersionTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleVersionTextBox.Location = new System.Drawing.Point(93, 0);
+            this.MetadataModuleVersionTextBox.Name = "MetadataModuleVersionLabel";
+            this.MetadataModuleVersionTextBox.Size = new System.Drawing.Size(250, 30);
+            this.MetadataModuleVersionTextBox.TabIndex = 2;
+            this.MetadataModuleVersionTextBox.Text = "0.0.0";
+            this.MetadataModuleVersionTextBox.ReadOnly = true;
+            this.MetadataModuleVersionTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataModuleVersionTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleVersionTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
-            // MetadataModuleLicenseLabel
+            // MetadataModuleLicenseTextBox
             //
-            this.MetadataModuleLicenseLabel.AutoSize = true;
-            this.MetadataModuleLicenseLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleLicenseLabel.Location = new System.Drawing.Point(93, 30);
-            this.MetadataModuleLicenseLabel.Name = "MetadataModuleLicenseLabel";
-            this.MetadataModuleLicenseLabel.Size = new System.Drawing.Size(250, 30);
-            this.MetadataModuleLicenseLabel.TabIndex = 4;
-            this.MetadataModuleLicenseLabel.Text = "None";
+            this.MetadataModuleLicenseTextBox.AutoSize = true;
+            this.MetadataModuleLicenseTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleLicenseTextBox.Location = new System.Drawing.Point(93, 30);
+            this.MetadataModuleLicenseTextBox.Name = "MetadataModuleLicenseLabel";
+            this.MetadataModuleLicenseTextBox.Size = new System.Drawing.Size(250, 30);
+            this.MetadataModuleLicenseTextBox.TabIndex = 4;
+            this.MetadataModuleLicenseTextBox.Text = "None";
+            this.MetadataModuleLicenseTextBox.ReadOnly = true;
+            this.MetadataModuleLicenseTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataModuleLicenseTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleLicenseTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
-            // MetadataModuleAuthorLabel
+            // MetadataModuleAuthorTextBox
             //
-            this.MetadataModuleAuthorLabel.AutoSize = true;
-            this.MetadataModuleAuthorLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleAuthorLabel.Location = new System.Drawing.Point(93, 60);
-            this.MetadataModuleAuthorLabel.Name = "MetadataModuleAuthorLabel";
-            this.MetadataModuleAuthorLabel.Size = new System.Drawing.Size(250, 30);
-            this.MetadataModuleAuthorLabel.TabIndex = 6;
-            this.MetadataModuleAuthorLabel.Text = "Nobody";
+            this.MetadataModuleAuthorTextBox.AutoSize = true;
+            this.MetadataModuleAuthorTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleAuthorTextBox.Location = new System.Drawing.Point(93, 60);
+            this.MetadataModuleAuthorTextBox.Name = "MetadataModuleAuthorTextBox";
+            this.MetadataModuleAuthorTextBox.Size = new System.Drawing.Size(250, 30);
+            this.MetadataModuleAuthorTextBox.TabIndex = 6;
+            this.MetadataModuleAuthorTextBox.Text = "Nobody";
+            this.MetadataModuleAuthorTextBox.ReadOnly = true;
+            this.MetadataModuleAuthorTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataModuleAuthorTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleAuthorTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // VersionLabel
             //
@@ -340,15 +371,19 @@
             this.VersionLabel.TabIndex = 1;
             this.VersionLabel.Text = "Version:";
             //
-            // MetadataModuleReleaseStatusLabel
+            // MetadataModuleReleaseStatusTextBox
             //
-            this.MetadataModuleReleaseStatusLabel.AutoSize = true;
-            this.MetadataModuleReleaseStatusLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleReleaseStatusLabel.Location = new System.Drawing.Point(93, 150);
-            this.MetadataModuleReleaseStatusLabel.Name = "MetadataModuleReleaseStatusLabel";
-            this.MetadataModuleReleaseStatusLabel.Size = new System.Drawing.Size(250, 30);
-            this.MetadataModuleReleaseStatusLabel.TabIndex = 11;
-            this.MetadataModuleReleaseStatusLabel.Text = "Stable";
+            this.MetadataModuleReleaseStatusTextBox.AutoSize = true;
+            this.MetadataModuleReleaseStatusTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleReleaseStatusTextBox.Location = new System.Drawing.Point(93, 150);
+            this.MetadataModuleReleaseStatusTextBox.Name = "MetadataModuleReleaseStatusLabel";
+            this.MetadataModuleReleaseStatusTextBox.Size = new System.Drawing.Size(250, 30);
+            this.MetadataModuleReleaseStatusTextBox.TabIndex = 11;
+            this.MetadataModuleReleaseStatusTextBox.Text = "Stable";
+            this.MetadataModuleReleaseStatusTextBox.ReadOnly = true;
+            this.MetadataModuleReleaseStatusTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataModuleReleaseStatusTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleReleaseStatusTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // MetadataModuleHomePageLinkLabel
             //
@@ -362,15 +397,19 @@
             this.MetadataModuleHomePageLinkLabel.Text = "linkLabel1";
             this.MetadataModuleHomePageLinkLabel.LinkClicked += new System.Windows.Forms.LinkLabelLinkClickedEventHandler(this.LinkLabel_LinkClicked);
             //
-            // MetadataModuleKSPCompatibilityLabel
+            // MetadataModuleKSPCompatibilityTextBox
             //
-            this.MetadataModuleKSPCompatibilityLabel.AutoSize = true;
-            this.MetadataModuleKSPCompatibilityLabel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.MetadataModuleKSPCompatibilityLabel.Location = new System.Drawing.Point(93, 180);
-            this.MetadataModuleKSPCompatibilityLabel.Name = "MetadataModuleKSPCompatibilityLabel";
-            this.MetadataModuleKSPCompatibilityLabel.Size = new System.Drawing.Size(250, 30);
-            this.MetadataModuleKSPCompatibilityLabel.TabIndex = 14;
-            this.MetadataModuleKSPCompatibilityLabel.Text = "0.0.0";
+            this.MetadataModuleKSPCompatibilityTextBox.AutoSize = true;
+            this.MetadataModuleKSPCompatibilityTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.MetadataModuleKSPCompatibilityTextBox.Location = new System.Drawing.Point(93, 180);
+            this.MetadataModuleKSPCompatibilityTextBox.Name = "MetadataModuleKSPCompatibilityLabel";
+            this.MetadataModuleKSPCompatibilityTextBox.Size = new System.Drawing.Size(250, 30);
+            this.MetadataModuleKSPCompatibilityTextBox.TabIndex = 14;
+            this.MetadataModuleKSPCompatibilityTextBox.Text = "0.0.0";
+            this.MetadataModuleKSPCompatibilityTextBox.ReadOnly = true;
+            this.MetadataModuleKSPCompatibilityTextBox.BackColor = MetadataTabPage.BackColor;
+            this.MetadataModuleKSPCompatibilityTextBox.ForeColor = System.Drawing.SystemColors.ControlText;
+            this.MetadataModuleKSPCompatibilityTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             //
             // MetadataModuleGitHubLinkLabel
             //
@@ -387,6 +426,16 @@
             // RelationshipTabPage
             //
             this.RelationshipTabPage.Controls.Add(this.DependsGraphTree);
+            this.RelationshipTabPage.Controls.Add(this.LegendDependsImage);
+            this.RelationshipTabPage.Controls.Add(this.LegendRecommendsImage);
+            this.RelationshipTabPage.Controls.Add(this.LegendSuggestsImage);
+            this.RelationshipTabPage.Controls.Add(this.LegendSupportsImage);
+            this.RelationshipTabPage.Controls.Add(this.LegendConflictsImage);
+            this.RelationshipTabPage.Controls.Add(this.LegendDependsLabel);
+            this.RelationshipTabPage.Controls.Add(this.LegendRecommendsLabel);
+            this.RelationshipTabPage.Controls.Add(this.LegendSuggestsLabel);
+            this.RelationshipTabPage.Controls.Add(this.LegendSupportsLabel);
+            this.RelationshipTabPage.Controls.Add(this.LegendConflictsLabel);
             this.RelationshipTabPage.Location = new System.Drawing.Point(4, 25);
             this.RelationshipTabPage.Name = "RelationshipTabPage";
             this.RelationshipTabPage.Padding = new System.Windows.Forms.Padding(3);
@@ -401,7 +450,7 @@
             | System.Windows.Forms.AnchorStyles.Left)
             | System.Windows.Forms.AnchorStyles.Right)));
             this.DependsGraphTree.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
-            this.DependsGraphTree.Location = new System.Drawing.Point(3, 7);
+            this.DependsGraphTree.Location = new System.Drawing.Point(3, 96);
             this.DependsGraphTree.Name = "DependsGraphTree";
             this.DependsGraphTree.Size = new System.Drawing.Size(345, 489);
             this.DependsGraphTree.TabIndex = 0;
@@ -418,6 +467,76 @@
             this.DependsGraphTree.ImageList.Images.Add("Suggests", global::CKAN.Properties.Resources.info);
             this.DependsGraphTree.ImageList.Images.Add("Supports", global::CKAN.Properties.Resources.smile);
             this.DependsGraphTree.ImageList.Images.Add("Conflicts", global::CKAN.Properties.Resources.alert);
+            //
+            // LegendDependsImage
+            //
+            this.LegendDependsImage.BackColor = System.Drawing.SystemColors.Window;
+            this.LegendDependsImage.Image = global::CKAN.Properties.Resources.star;
+            this.LegendDependsImage.Location = new System.Drawing.Point(6, 3);
+            this.LegendDependsImage.SizeMode = System.Windows.Forms.PictureBoxSizeMode.StretchImage;
+            this.LegendDependsImage.ClientSize = new System.Drawing.Size(14, 14);
+            //
+            // LegendRecommendsImage
+            //
+            this.LegendRecommendsImage.BackColor = System.Drawing.SystemColors.Window;
+            this.LegendRecommendsImage.Image = global::CKAN.Properties.Resources.thumbup;
+            this.LegendRecommendsImage.Location = new System.Drawing.Point(6, 21);
+            this.LegendRecommendsImage.SizeMode = System.Windows.Forms.PictureBoxSizeMode.StretchImage;
+            this.LegendRecommendsImage.ClientSize = new System.Drawing.Size(14, 14);
+            //
+            // LegendSuggestsImage
+            //
+            this.LegendSuggestsImage.BackColor = System.Drawing.SystemColors.Window;
+            this.LegendSuggestsImage.Image = global::CKAN.Properties.Resources.info;
+            this.LegendSuggestsImage.Location = new System.Drawing.Point(6, 39);
+            this.LegendSuggestsImage.SizeMode = System.Windows.Forms.PictureBoxSizeMode.StretchImage;
+            this.LegendSuggestsImage.ClientSize = new System.Drawing.Size(14, 14);
+            //
+            // LegendSupportsImage
+            //
+            this.LegendSupportsImage.BackColor = System.Drawing.SystemColors.Window;
+            this.LegendSupportsImage.Image = global::CKAN.Properties.Resources.smile;
+            this.LegendSupportsImage.Location = new System.Drawing.Point(6, 57);
+            this.LegendSupportsImage.SizeMode = System.Windows.Forms.PictureBoxSizeMode.StretchImage;
+            this.LegendSupportsImage.ClientSize = new System.Drawing.Size(14, 14);
+            //
+            // LegendConflictsImage
+            //
+            this.LegendConflictsImage.BackColor = System.Drawing.SystemColors.Window;
+            this.LegendConflictsImage.Image = global::CKAN.Properties.Resources.alert;
+            this.LegendConflictsImage.Location = new System.Drawing.Point(6, 75);
+            this.LegendConflictsImage.SizeMode = System.Windows.Forms.PictureBoxSizeMode.StretchImage;
+            this.LegendConflictsImage.ClientSize = new System.Drawing.Size(14, 14);
+            //
+            // LegendDependsLabel
+            //
+            this.LegendDependsLabel.AutoSize = true;
+            this.LegendDependsLabel.Location = new System.Drawing.Point(24, 3);
+            this.LegendDependsLabel.Text = "Depends";
+            //
+            // LegendRecommendsLabel
+            //
+            this.LegendRecommendsLabel.AutoSize = true;
+            this.LegendRecommendsLabel.Location = new System.Drawing.Point(24, 21);
+            this.LegendRecommendsLabel.Text = "Recommends";
+            //
+            // LegendSuggestsLabel
+            //
+            this.LegendSuggestsLabel.AutoSize = true;
+            this.LegendSuggestsLabel.Location = new System.Drawing.Point(24, 39);
+            this.LegendSuggestsLabel.Text = "Suggests";
+            //
+            // LegendSupportsLabel
+            //
+            this.LegendSupportsLabel.AutoSize = true;
+            this.LegendSupportsLabel.Location = new System.Drawing.Point(24, 57);
+            this.LegendSupportsLabel.Text = "Supports";
+            //
+            // LegendConflictsLabel
+            //
+            this.LegendConflictsLabel.AutoSize = true;
+            this.LegendConflictsLabel.Location = new System.Drawing.Point(24, 75);
+            this.LegendConflictsLabel.Text = "Conflicts";
             //
             // ContentTabPage
             //
@@ -513,28 +632,38 @@
         private System.Windows.Forms.TabPage MetadataTabPage;
         private System.Windows.Forms.SplitContainer splitContainer2;
         private System.Windows.Forms.TableLayoutPanel MetaDataUpperLayoutPanel;
-        private System.Windows.Forms.Label MetadataModuleNameLabel;
+        private TransparentTextBox MetadataModuleNameTextBox;
         private System.Windows.Forms.Label MetadataModuleAbstractLabel;
-        private System.Windows.Forms.Label MetadataModuleDescriptionLabel;
+        private TransparentTextBox MetadataModuleDescriptionTextBox;
         private System.Windows.Forms.TableLayoutPanel MetaDataLowerLayoutPanel;
         private System.Windows.Forms.Label IdentifierLabel;
-        private System.Windows.Forms.Label MetadataIdentifierLabel;
+        private TransparentTextBox MetadataIdentifierTextBox;
         private System.Windows.Forms.Label KSPCompatibilityLabel;
         private System.Windows.Forms.Label ReleaseLabel;
         private System.Windows.Forms.Label GitHubLabel;
         private System.Windows.Forms.Label HomePageLabel;
         private System.Windows.Forms.Label AuthorLabel;
         private System.Windows.Forms.Label LicenseLabel;
-        private System.Windows.Forms.Label MetadataModuleVersionLabel;
-        private System.Windows.Forms.Label MetadataModuleLicenseLabel;
-        private System.Windows.Forms.Label MetadataModuleAuthorLabel;
+        private TransparentTextBox MetadataModuleVersionTextBox;
+        private TransparentTextBox MetadataModuleLicenseTextBox;
+        private TransparentTextBox MetadataModuleAuthorTextBox;
         private System.Windows.Forms.Label VersionLabel;
-        private System.Windows.Forms.Label MetadataModuleReleaseStatusLabel;
+        private TransparentTextBox MetadataModuleReleaseStatusTextBox;
         private System.Windows.Forms.LinkLabel MetadataModuleHomePageLinkLabel;
-        private System.Windows.Forms.Label MetadataModuleKSPCompatibilityLabel;
+        private TransparentTextBox MetadataModuleKSPCompatibilityTextBox;
         private System.Windows.Forms.LinkLabel MetadataModuleGitHubLinkLabel;
         private System.Windows.Forms.TabPage RelationshipTabPage;
         private System.Windows.Forms.TreeView DependsGraphTree;
+        private System.Windows.Forms.PictureBox LegendDependsImage;
+        private System.Windows.Forms.PictureBox LegendRecommendsImage;
+        private System.Windows.Forms.PictureBox LegendSuggestsImage;
+        private System.Windows.Forms.PictureBox LegendSupportsImage;
+        private System.Windows.Forms.PictureBox LegendConflictsImage;
+        private System.Windows.Forms.Label LegendDependsLabel;
+        private System.Windows.Forms.Label LegendRecommendsLabel;
+        private System.Windows.Forms.Label LegendSuggestsLabel;
+        private System.Windows.Forms.Label LegendSupportsLabel;
+        private System.Windows.Forms.Label LegendConflictsLabel;
         private System.Windows.Forms.TabPage ContentTabPage;
         private System.Windows.Forms.TreeView ContentsPreviewTree;
         private System.Windows.Forms.Button ContentsDownloadButton;